# src/swagger.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Merchant Backend API
  description: PumaPay API that allows the merchants through the Merchant Backend to create and execute crypto payments using PumaPay Pull PullPaymentModel Protocol over the Ethereum network.
  contact:
    name: PumaPay Developers Team
    email: developers@pumapay.io
    url: http://swagger.io
servers:
  - url: http://localhost:3000/
    description: Local Server
  - url: http://stgmbackend.pumapay.io/merchant/
    description: Test Server
paths:
...
  /api/v2/pull-payments/:
    post:
      tags:
        - Pull Payments
      summary: 'Creates a new pull payment'
      parameters: []
      description: 'Creates a new pull payment in the merchant database'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/PullPaymentRequest'
      responses:
        '200':
          description: Returns the newly created pull payment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessHttpResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        $ref: '#/components/schemas/PullPayment'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorHttpResponse'
                  - type: object
                    properties:
                      error:
                        type: string
    get:
      tags:
        - Pull Payments
      summary: 'Retrieve all the pull payments'
      parameters: []
      responses:
        '200':
          description: Returns all the pull payments from the merchant DB
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessHttpResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          $ref: '#/components/schemas/PullPayment'
        '400':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorHttpResponse'
  /api/v2/pull-payments/{PullPaymentID}:
    get:
      tags:
        - Pull Payments
      summary: 'Retrieve the pull payment by pull payment ID'
      parameters:
        - in: path
          name: PullPaymentID
          schema:
            type: string
          required: true
          description: ID of the pull payment to retrieve
      description: 'Retrieve the pull payment by ID that exists in the merchant database'
      responses:
        '200':
          description: Returns the payment by ID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessHttpResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        $ref: '#/components/schemas/PullPayment'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorHttpResponse'
                  - type: object
                    properties:
                      error:
                        type: string
    delete:
      tags:
        - Pull Payments
      summary: 'Deletes a pull payment'
      parameters:
        - in: path
          name: PullPaymentID
          schema:
            type: string
          required: true
          description: ID of the pull payment to delete
      description: 'Deletes a pull payment from the merchant database'
      responses:
        '200':
          description: Deletes a pull payment from the merchant database
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessHttpResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorHttpResponse'
                  - type: object
                    properties:
                      error:
                        type: string
...
components:
 securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: fcm-mobile-token
    ApiKey:
      description: 'Api Key'
      type: 'apiKey'
      in: 'header'
      name: 'pma-api-key'
 schemas:
    SuccessHttpResponse:
      required:
        - code
        - success
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Descriptive Success message
        data:
          type: object
    ErrorHttpResponse:
      required:
        - code
        - success
        - message
        - error
      properties:
        code:
          type: integer
          example: 400
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
...
    PullPayment:
      type: object
      properties:
        id:
          type: string
          example: 4a17335e-bf18-11e8-a355-222222fb1459
          format: uuid
          description: Auto-generated ID for the pull payment
        hdWalletIndex:
          type: integer
          example: 1
          format: integer
          description: The wallet index attached with this pull payment
        pullPaymentModelID:
          type: string
          example: 4a17335e-bf18-11e8-a355-fb1459
          format: uuid
          description: Pull Payment model ID linked with this pull payment
        numberOfPayments:
          type: integer
          example: 12
          format: integer
          description: Number of payments that the merchant will be allowed to pull from the customer's wallet
        nextPaymentDate:
          type: string
          example: 1537747200
          format: unix-epoch-in-seconds
          description: Unix Epoch Timestamp of the next pull payment
        lastPaymentDate:
          type: string
          example: 1537401600
          format: unix-epoch-in-seconds
          description: Unix Epoch Timestamp of the last pull payment
        startTimestamp:
          type: string
          example: 1537401600
          format: unix-epoch-in-seconds
          description: Unix Epoch Timestamp of the beginning of the pull payment
        customerAddress:
          type: string
          example: '0x0000000000000000000000000000000000011111'
          format: eth-address
          description: The ethereum address of the customer
        merchantAddress:
          type: string
          example: '0x0000000000000000000000000000000000022222'
          format: eth-address
          description: The ethereum address of the merchant
        pullPaymentAddress:
          type: string
          example: '0x0000000000000000000000000000000000033333'
          format: eth-address
          description: The ethereum address of the pull payment smart contract
        statusID:
          type: integer
          enum: [initial = 1, running = 2, stopped = 3, cancelled = 4, done = 5]
          example: 2
          format: integer
          description: Payment Status ID
        userID:
          type: string
          example: userIdOnMerchantsSystem
          description: The userID of the customer as it is being used in merchant's existing system
    PullPaymentRequest:
      type: object
      required:
        - PullPaymentModelID
        - numberOfPayments
        - startTimestamp
        - customerAddress
      properties:
        pullPaymentModelID:
          type: string
          example: 4a17335e-bf18-11e8-a355-fb1459
          format: uuid
          description: Pull Payment model ID linked with this pull payment
        numberOfPayments:
          type: integer
          example: 12
          format: integer
          description: Number of payments that the merchant will be allowed to pull from the customer's wallet
        startTimestamp:
          type: string
          example: 1537401600
          format: unix-epoch-in-seconds
          description: Unix Epoch Timestamp of the beginning of the pull payment
        customerAddress:
          type: string
          example: '0x0000000000000000000000000000000000011111'
          format: eth-address
          description: The ethereum address of the customer
        pullPaymentAddress:
          type: string
          example: '0x0000000000000000000000000000000000033333'
          format: eth-address
          description: The ethereum address of the pull payment smart contract
        userID:
          type: string
          example: userIdOnMerchantsSystem
          description: The userID of the customer as it is being used in merchant's existing system
...